Move Git helper to separate file
